Add FAQ section to buyers page
